Fix typo to fit with Android

+++ src/megachatapi.h
@@ -612,13 +612,13 @@ public:
     static constexpr int CALL_LIMIT_DURATION_DISABLED = -1;
 
     // Call limit restriction no present (Call limit won't be modified)
-    static constexpr uint32_t CALL_LIMIT_NO_PRESENT = UINT32_MAX;
+    static constexpr unsigned long CALL_LIMIT_NO_PRESENT = UINT32_MAX;
 
     // Maximum number of clients with which a single user can join a call
-    static constexpr uint32_t CALL_LIMIT_USERS_PER_CLIENT = 4;
+    static constexpr unsigned long CALL_LIMIT_USERS_PER_CLIENT = 4;
 
     // No limit set for a call option like (duration, max clients per call...)
-    static constexpr uint32_t CALL_LIMIT_RESET = 0;
+    static constexpr unsigned long CALL_LIMIT_RESET = 0;
 
     virtual ~MegaChatCall();
